requirements.txt
boto3>=1.9.86
plaid-python==2.4.0 
requests>=2.21.0
